+++ cli/NewUserCommand.php
@@ -12,6 +12,7 @@ use Grav\Console\ConsoleCommand;
 use Grav\Common\File\CompiledYamlFile;
 use Grav\Common\User\User;
 use Grav\Common\Grav;
+use Grav\Plugin\Login\Login;
 use RocketTheme\Toolbox\ResourceLocator\UniformResourceLocator;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Helper\Helper;
@@ -25,11 +26,11 @@ use Symfony\Component\Console\Question\Question;
  */
 class NewUserCommand extends ConsoleCommand
 {
-
-    /**
-     * @var array
-     */
+    /** @var array */
     protected $options = [];
+
+    /** @var Login */
+    protected $login;
 
     /**
      * Configure the command
@@ -91,6 +92,14 @@ class NewUserCommand extends ConsoleCommand
      */
     protected function serve()
     {
+        include __DIR__ . '/../vendor/autoload.php';
+
+        $grav = Grav::instance();
+        if (!isset($grav['login'])) {
+            $grav['login'] = new Login($grav);
+        }
+        $this->login = $grav['login'];
+
         $this->options = [
             'user'        => $this->input->getOption('user'),
             'password1'   => $this->input->getOption('password'),
@@ -113,7 +122,13 @@ class NewUserCommand extends ConsoleCommand
             // Get username and validate
             $question = new Question('Enter a <yellow>username</yellow>: ', 'admin');
             $question->setValidator(function ($value) {
-                return $this->validate('user', $value);
+                $this->validate('user', $value);
+
+                if (User::find($value, ['username'])->exists()) {
+                    throw new \RuntimeException('Username "' . $value . '" already exists, please pick another username');
+                };
+
+                return true;
             });
 
             $username = $helper->ask($this->input, $this->output, $question);
@@ -238,79 +253,15 @@ class NewUserCommand extends ConsoleCommand
     }
 
     /**
-     * @param        $type
-     * @param        $value
+     * @param string $type
+     * @param mixed  $value
      * @param string $extra
      *
-     * @return mixed
+     * @return string
      */
     protected function validate($type, $value, $extra = '')
     {
-        /** @var Config $config */
-        $config = Grav::instance()['config'];
-
-        /** @var UniformResourceLocator $locator */
-        $locator = Grav::instance()['locator'];
-
-
-        $username_regex = '/' . $config->get('system.username_regex') . '/';
-        $pwd_regex      = '/' . $config->get('system.pwd_regex') . '/';
-
-        switch ($type) {
-            case 'user':
-                if (!preg_match($username_regex, $value)) {
-                    throw new \RuntimeException('Username should be between 3 and 16 characters, including lowercase letters, numbers, underscores, and hyphens. Uppercase letters, spaces, and special characters are not allowed');
-                }
-                if (file_exists($locator->findResource('account://' . $value . YAML_EXT))) {
-                    throw new \RuntimeException('Username "' . $value . '" already exists, please pick another username');
-                }
-
-                break;
-
-            case 'password1':
-                if (!preg_match($pwd_regex, $value)) {
-                    throw new \RuntimeException('Password must contain at least one number and one uppercase and lowercase letter, and at least 8 or more characters');
-                }
-
-                break;
-
-            case 'password2':
-                if (strcmp($value, $extra)) {
-                    throw new \RuntimeException('Passwords did not match.');
-                }
-
-                break;
-
-            case 'email':
-                if (!filter_var($value, FILTER_VALIDATE_EMAIL)) {
-                    throw new \RuntimeException('Not a valid email address');
-                }
-
-                break;
-
-            case 'permissions':
-                if (!in_array($value, ['a', 's', 'b'])) {
-                    throw new \RuntimeException('Permissions ' . $value . ' are invalid.');
-                }
-
-                break;
-
-            case 'fullname':
-                if ($value === null || trim($value) === '') {
-                    throw new \RuntimeException('Fullname cannot be empty');
-                }
-
-                break;
-
-            case 'state':
-                if ($value !== 'enabled' && $value !== 'disabled') {
-                    throw new \RuntimeException('State is not valid');
-                }
-
-                break;
-        }
-
-        return $value;
+        return $this->login->validateField($type, $value, $extra);
     }
 
     /**