+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+# v2.7.0
+## 05/11/2018
+
+1. [](#new)
+    * Moved support for 2FA authentication into Login plugin (only supported in Admin currently)
+    * Updated plugin dependencies (Grav >= 1.4.5, Form >=2.13.4, Email >=2.7.0)
+1. [](#improved)
+    * Added cleaner way for 3rd party providers to add twig templates to login form
+    * Use `Login` class validation methods in CLI
+    * Added logging of login exceptions
+    * Show denied message only when authenticated but not authorized
+1. [](#bugfix)
+    * Don't allow Profile saving if a Grav user account doesn't exist (OAuth/LDAP users for example)
+    * Don't allow PW reset if no current password exists (OAuth/LDAP users for example) 
+
 # v2.6.3
 ## 04/12/2018